+++ cooked_input/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-__version__ = '0.2.2'
+__version__ = '0.2.3'
 
 from .get_input import get_input, get_table_input, process_value, make_pretty_table
 from .get_input import get_string, get_int, get_float, get_boolean, get_date, get_yes_no, get_list
@@ -15,6 +15,6 @@ from .validators import InAnyValidator, NotInValidator, InChoicesValidator, Rege
 from .validators import ListValidator, SimpleValidator
 from .validators import in_all, in_any, not_in, validate
 from .cleaners import Cleaner, UpperCleaner, LowerCleaner, CapitalizeCleaner
-from .cleaners import StripCleaner, ReplaceCleaner
+from .cleaners import StripCleaner, ReplaceCleaner, ChoiceCleaner, RegexCleaner
 
 __all__ = ['get_input', 'convertors', 'validators', 'cleaners', 'error_callbacks']